fix: check for old messages at DeleteVideoUsecase.ts

+++ src/usecases/DeleteVideoUsecase.ts
@@ -1,5 +1,6 @@
 import { args } from '@/config'
 import { type Preset, Usecase } from '@/domain'
+import { ImplementationError } from '@/errors'
 import { VideosRepository, VideoUploadsRepository } from '@/repositories'
 import { type CLIService, TelegramService } from '@/services'
 
@@ -56,6 +57,38 @@ export class DeleteVideoUsecase extends Usecase {
       return 'MENU'
     }
 
+    const videoUploads = await this.videoUploadsRepository.getAll(selectedVideo.id)
+
+    const [firstVideoUpload] = videoUploads
+    if (!firstVideoUpload) {
+      throw new ImplementationError('Undefined firstVideoUpload')
+    }
+
+    const twoDaysInMilliseconds = 1_000 * 60 * 60 * 24 * 2
+
+    const isUploadOlderThanTwoDays =
+      firstVideoUpload.uploadedAt.getTime() + twoDaysInMilliseconds <= Date.now()
+
+    if (isUploadOlderThanTwoDays) {
+      this.cliService.printError(
+        `Telegram doesn't allow to delete a message older than two days through API`
+      )
+
+      const shouldDeleteOnlyFromDatabase = await this.cliService.confirm({
+        message: 'Delete ONLY from database?',
+        default: false
+      })
+
+      if (!shouldDeleteOnlyFromDatabase) {
+        const shouldReturnToMenu = await this.cliService.confirm({
+          message: 'Return to menu?',
+          default: true
+        })
+
+        return shouldReturnToMenu ? 'MENU' : 'OK'
+      }
+    }
+
     if (args.dryRun) {
       this.printDryRunMessage()
       return 'OK'
@@ -68,16 +101,27 @@ export class DeleteVideoUsecase extends Usecase {
 
     deleteLoading.start()
 
-    const videoUploads = await this.videoUploadsRepository.getAll(selectedVideo.id)
-
-    await this.telegramService.deleteMessages({
-      channelId: this.preset.telegram.channelId,
-      messagesIds: videoUploads.map(videoUpload => videoUpload.telegramPostId)
-    })
+    if (!isUploadOlderThanTwoDays) {
+      await this.telegramService.deleteMessages({
+        channelId: this.preset.telegram.channelId,
+        messagesIds: videoUploads.map(videoUpload => videoUpload.telegramPostId)
+      })
+    }
 
     await this.videosRepository.deleteFromId(selectedVideo.id)
 
-    deleteLoading.stop()
+    deleteLoading.stop(
+      isUploadOlderThanTwoDays ? 'Successfully deleted ONLY FROM DATABASE.' : undefined
+    )
+
+    if (isUploadOlderThanTwoDays) {
+      const oldMessageInfo = [
+        'Keep in mind that the desynchronized messages in Telegram are still there.',
+        'If you still want to remove them, you should do that manually.'
+      ].join('\n')
+
+      this.cliService.printInfo(oldMessageInfo)
+    }
 
     const shouldReturnToMenu = await this.cliService.confirm({
       message: 'Return to menu?',
